Add files via upload
Feedback Riccardo
Good work. Some minor issues:

"The pp-check seems to generate data that is similar to our simulated data, with some slight variations - which makes sense given the priors are weakly informed." Could be better expressed.  Variations between blue lines are due to the model being probabilistic (instead of deterministic), variations btw black and blue are due to weakly informed/skeptical priors centered at 0, regardless of evidence in the data for effects in one direction (since the priors here are a priori).
"the model is accepted as a good fit and will be used when introducing collected data in the next question". I'd still like you to comment as to what pub bias does to estimates
sampling variance: the formula you use is correct for normal sd. However, it's a bit misleading here, since we are talking about pooled sd, which has a different formula (see slides)
"it can be concluded that there is a publication bias within this sample of the field of study of pitch of patients with schizophrenia " not really. We can suggest, not really conclude. Also I'd need you to comment on estimates from the mode and what pub bias would imply for how to interpret them